+++ src/test/java/roomescape/model/ReservationTest.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.model;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
